+++ CHANGELOG.md
@@ -17,6 +17,10 @@ For a list of breaking changes, check [here](#breaking-changes).
 <!-- - [ ] update lein-clj-kondo -->
 <!-- - [ ] update carve -->
 
+## Unreleased
+
+- [#2741](https://github.com/clj-kondo/clj-kondo/issues/2741): New linter `:duplicate-require-option` which warns on duplicate `:exclude` clauses in `:require` and `:refer-clojure` options and consolidates the previously separate `:duplicate-refer` linter into this one. ([@jramosg](https://github.com/jramosg))
+
 ## 2026.01.19
 
 - [#2735](https://github.com/clj-kondo/clj-kondo/issues/2735): NEW linter: `:duplicate-refer` which warns on duplicate entries in `:refer` of `:require`. ([@jramosg](https://github.com/jramosg))